fix(B-464): split user_profile model and schema
